try laczos-arpack

+++ nonlinear-rotors/codes-lanczos/submission-lanc.py
@@ -19,7 +19,7 @@ def jobstring(NameOfServer,Rpt,jmax,dir_output,niter):
 	job_string = """#!/bin/bash
 #SBATCH --job-name=%s
 #SBATCH --output=%s
-#SBATCH --time=1-00:00
+#SBATCH --time=2-00:00
 %s
 #SBATCH --mem-per-cpu=8GB
 #SBATCH --cpus-per-task=1
@@ -29,9 +29,9 @@ export OMP_NUM_THREADS=1
 	return job_string
 
 #initial parameters for qmc.input
-status = 'S'
-niter = 50
-jrot = 5
+status = 'A'
+niter = 400
+jrot = 2
 zmin = 2.6
 zmax = 10.0
 dz = 0.1
@@ -49,11 +49,11 @@ if (status == 'S'):
 	print(run_command)
 
 if (status == 'A'):
-	eigvalvsRpt1 = np.zeros(nz)
-	eigvalvsRpt2 = np.zeros(nz)
-	saved_Rpt = np.zeros(nz)
-	SvNvsRpt = np.zeros(nz)
-	S2vsRpt = np.zeros(nz)
+	eigval=np.zeros((nz,3))
+	err=np.zeros((nz,3))
+	saved_Rpt=np.zeros(nz)
+	SvNvsRpt=np.zeros(nz)
+	S2vsRpt=np.zeros(nz)
 
 # Loop over your jobs
 index_end=0
@@ -66,52 +66,76 @@ for i in range(nz):
 	if (status == 'S'):
 		os.chdir(dir_output)
 		#job submission
-		fname = 'lanc_Rpt'+str(Rpt)+"Ang"
-		fwrite = open(fname, 'w')
+		fname = 'lanc_Rpt'+str(Rpt)+"Ang-J"+str(jrot)+"-niter"+str(niter)
+		if (os.path.isfile(fname) == True):
+			print("job name - "+fname)
+			print("It has already been submitted")
+			os.chdir(src_dir)
+			exit()
+		else:
+			fwrite = open(fname, 'w')
 
-		fwrite.write(jobstring(NameOfServer,Rpt,jrot,dir_output,niter))
-		fwrite.close()
-		#call(["sbatch", "-p", "highmem", fname])
-		call(["sbatch", fname])
+			fwrite.write(jobstring(NameOfServer,Rpt,jrot,dir_output,niter))
+			fwrite.close()
+			#call(["sbatch", "-p", "highmem", fname])
+			call(["sbatch", fname])
+			#call(["sbatch", "-C", "faster", fname])
 
 		os.chdir(src_dir)
 
 	if (status == "A"):
 		saved_Rpt[i] = Rpt
-		thetaNum = int(2*jrot+3)
-		angleNum = int(2*(2*jrot+1))
+		if (jrot <= 6):
+			thetaNum = int(2*jrot+3)
+			angleNum = int(2*(2*jrot+1))
+		else:
+			thetaNum = int(jrot+2)
+			angleNum = int(2*jrot+2)
+			
 		strFile = "lanc-2-p-H2O-jmax"+str(jrot)+"-Rpt"+str(Rpt)+"Angstrom-grid-"+str(thetaNum)+"-"+str(angleNum)+"-niter"+str(niter)+".txt"
 
-		fileAnalyze_energy = "ground-state-energy-"+strFile
+		#fileAnalyze_energy = "ground-state-energy-"+strFile
+		fileAnalyze_energy = "energy-levels-"+strFile
 		data_input_energy = dir_output+"/"+fileAnalyze_energy
 		CMRECIP2KL = 1.4387672
-		if (os.path.isfile(data_input_energy) == True):
-			eig_kelvin= np.loadtxt(data_input_energy, usecols=(0), unpack=True)
-			eigvalvsRpt1[i] = eig_kelvin
-			eigvalvsRpt2[i] = eig_kelvin/CMRECIP2KL
-			index_end = index_end+1
-			print(index_end)
-		else:
-			break
 
-		'''
 		fileAnalyze_entropy = "ground-state-entropies-"+strFile
 		data_input_entropy = dir_output+"/"+fileAnalyze_entropy
-		ent = np.loadtxt(data_input_entropy, usecols=(0), unpack=True)
-		SvNvsRpt[i] = ent[0]
-		S2vsRpt[i] = ent[1]
-		'''
+
+		if (os.path.isfile(data_input_energy) == True):
+			print(index_end)
+			print(data_input_energy)
+			eig_kelvin, err_kelvin = np.loadtxt(data_input_energy, usecols=(0, 1), unpack=True)
+			for j in range(np.size(eigval,1)):
+				eigval[i,j] = eig_kelvin[j]
+				err[i,j] = err_kelvin[j]
+				print(eigval[i,j])
+			#eigvalvsRpt2[i] = eig_kelvin/CMRECIP2KL
+		else:
+			pass
+
+		index_end = index_end+1
+
+		if (os.path.isfile(data_input_entropy) == True):
+			ent = np.loadtxt(data_input_entropy, usecols=(0), unpack=True)
+			SvNvsRpt[i] = ent[0]
+			S2vsRpt[i] = ent[1]
+		else:
+			pass
+
 
 if (status == "A"):
 	#printing block is opened
 	strFile1 = "lanc-2-p-H2O-jmax"+str(jrot)+"-grid-"+str(thetaNum)+"-"+str(angleNum)+"-niter"+str(niter)+".txt"
 
-	energy_comb = np.array([saved_Rpt[:index_end], eigvalvsRpt1[:index_end], eigvalvsRpt2[:index_end]])
+	energy_comb1=np.concatenate((eigval[:index_end,:], err[:index_end,:]), axis=1)
+	energy_comb=np.concatenate((saved_Rpt[:index_end,np.newaxis],energy_comb1), axis=1)
 	eig_file = dir_output+"/ground-state-energy-vs-Rpt-"+strFile1
-	np.savetxt(eig_file, energy_comb.T, fmt='%20.8f', delimiter=' ', header='First col. --> Rpt (Angstrom); 2nd and 3rd cols are the eigen values of (Htot = Hrot + Hvpot) in Kelvin and wavenumber, respectively. ')
+	np.savetxt(eig_file, energy_comb, fmt='%1.5e', delimiter='    ', header='col1-> Rpt (Angstrom); col2->ground state energy in kelvin; col3->first excited state energy in kelvin; col4->second excited state energy in kelvin. The columns 5-7 are the corresponding errors.')
 
-	#entropy_comb = np.array([saved_Rpt, SvNvsRpt, S2vsRpt])
-	#ent_file = dir_output+"/ground-state-entropies-vs-Rpt-"+strFile1
-	#np.savetxt(ent_file, entropy_comb.T, fmt='%20.8f', delimiter=' ', header='First col. --> Rpt (Angstrom); 2nd and 3rd cols are the S_vN and S_2, respectively. ')
+	#entropy_comb = np.array([saved_Rpt[:index_end-1], SvNvsRpt[:index_end-1], S2vsRpt[:index_end-1]])
+	entropy_comb = np.array([saved_Rpt[:index_end], SvNvsRpt[:index_end], S2vsRpt[:index_end]])
+	ent_file = dir_output+"/ground-state-entropies-vs-Rpt-"+strFile1
+	np.savetxt(ent_file, entropy_comb.T, fmt='%20.8f', delimiter=' ', header='First col. --> Rpt (Angstrom); 2nd and 3rd cols are the S_vN and S_2, respectively. ')
 	# printing block is closed
